fix: accepting pointer to empty struct directly now
The update function was using reflect to determine the object type and
creating new pointer. First of all, it was actually returning the
pointer to pointer to the object which does not implement
interfaces.Object so all reflect calls were failing.

I tried to fix that with Elem call but that returns the type with zero
value and since the type is pointer its zero value is nil. However
client requires pointer to empty struct not nil pointer. I have not
figured out a way so I added the comment. For now I'm accepting the
ExistingObject empty struct directly and using it.